+++ examples/hello.rs
@@ -1,12 +1,13 @@
 //! A minimal example with a terminal and camera.
 
-use bevy::{color::palettes::css, prelude::*};
+use bevy::{color::palettes::css, prelude::*, window::WindowMode};
 use bevy_ascii_terminal::*;
 
 fn main() {
     App::new()
         .add_plugins((DefaultPlugins, TerminalPlugins))
         .add_systems(Startup, setup)
+        .add_systems(Update, handle_input)
         .run();
 }
 
@@ -14,7 +15,6 @@ fn setup(mut commands: Commands) {
     commands.spawn(
         Terminal::new([14, 3])
             .with_border(BoxStyle::SINGLE_LINE)
-            .with_padding(Padding::ONE)
             .with_bg_clear_color(css::DARK_BLUE.into())
             .with_string(
                 [0, 0],
@@ -22,4 +22,21 @@ fn setup(mut commands: Commands) {
             ),
     );
     commands.spawn(TerminalCamera::new());
+}
+
+fn handle_input(
+    input: Res<ButtonInput<KeyCode>>,
+    mut win: Single<&mut Window>,
+    mut exit: MessageWriter<AppExit>,
+) {
+    if input.just_pressed(KeyCode::Escape) {
+        exit.write(AppExit::Success);
+    }
+    if input.just_pressed(KeyCode::KeyF) {
+        if win.mode == WindowMode::BorderlessFullscreen(MonitorSelection::Current) {
+            win.mode = WindowMode::Windowed;
+        } else {
+            win.mode = WindowMode::BorderlessFullscreen(MonitorSelection::Current);
+        }
+    }
 }