+++ examples/rexpaint.rs
@@ -1,12 +1,13 @@
 //! A terminal built from a rexpaint file.
 
-use bevy::prelude::*;
+use bevy::{prelude::*, window::WindowMode};
 use bevy_ascii_terminal::*;
 
 fn main() {
     App::new()
         .add_plugins((DefaultPlugins, TerminalPlugins))
         .add_systems(Startup, setup)
+        .add_systems(Update, handle_input)
         .run();
 }
 
@@ -15,4 +16,21 @@ fn setup(mut commands: Commands) {
         .unwrap()
         .with_border(BoxStyle::SINGLE_LINE),));
     commands.spawn(TerminalCamera::new());
+}
+
+fn handle_input(
+    input: Res<ButtonInput<KeyCode>>,
+    mut win: Single<&mut Window>,
+    mut exit: MessageWriter<AppExit>,
+) {
+    if input.just_pressed(KeyCode::Escape) {
+        exit.write(AppExit::Success);
+    }
+    if input.just_pressed(KeyCode::KeyF) {
+        if win.mode == WindowMode::BorderlessFullscreen(MonitorSelection::Current) {
+            win.mode = WindowMode::Windowed;
+        } else {
+            win.mode = WindowMode::BorderlessFullscreen(MonitorSelection::Current);
+        }
+    }
 }