+++ src/strings/mod.rs
@@ -1,8 +1,8 @@
-mod parse;
+mod tagged_string;
 mod terminal_string;
 
 pub use terminal_string::{TerminalString, TerminalStringBuilder};
 
-pub use parse::{
+pub use tagged_string::{
     Token, TokenIterator, wrap_line_count, wrap_string, wrap_tagged_line_count, wrap_tagged_string,
 };